Refactor environment variable handling and improve API error handling
- Removed environment variables from next.config.js for better security practices.
- Updated package.json to use yarn for lint:fix command.
- Enhanced error handling in API routes to provide more detailed error messages and logging.
- Added new API routes for Polkadot proxy and provider configuration.
- Improved data validation and error handling in existing API functions.
- Updated chart components to include loading states and error boundaries for better user experience.
- Added explorer links and pool address properties to pool data structure.
- Updated README.md with new environment variable instructions and API usage details.

+++ src/components/wallet/UnstakingDisplay.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { Popover, PopoverButton, PopoverPanel } from '@headlessui/react';
 import { IoMdHelpCircleOutline } from "react-icons/io";
 import { abbreviateNumber } from '../../utils/helpers';
-import { BTreeMap, u128, u64 } from "@polkadot/types-codec";
+import { BTreeMap, u128, u32,u64 } from "@polkadot/types-codec";
 import { StakingService, SystemInfoService } from "@curiodao/capital-dex-sdk/polkadot";
 import { useWallet } from '@/constant/context/WalletContext';
 import { toast } from 'react-toastify';
@@ -11,7 +11,7 @@ interface UnstakingDisplayProps {
   totalStaked: string;
   marketPrice: number;
   unstaking: BTreeMap<u64, u128> | undefined;
-  currentBlock: u64 | undefined;
+  currentBlock: u32 | undefined;
   updateWalletInfo: () => void;
 }
